isic_2019/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.4
Jinja2==2.10
lime==0.1.1.36
matplotlib==3.0.3
numpy==1.15.4
onnx==1.2.1
pandas==0.22.0
Pillow==5.2.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
sklearn==0.0
tornado==5.0.2
urllib3==1.23
WTForms==2.2.1
